Move MockDaService to state_machine/mocks
