+++ rollup-interface/src/state_machine/mocks/mod.rs
@@ -1,9 +1,12 @@
 //! Defines mock instantiations of many important traits, which are useful
 //! for testing, fuzzing, and benchmarking.
 
+mod da;
 mod validity_condition;
 mod zk_vm;
-mod da;
+pub use da::{
+    MockAddress, MockBatchBuilder, MockDaService, MockDaSpec, TestBlob, TestBlock, TestBlockHeader,
+    TestHash,
+};
 pub use validity_condition::TestValidityCond;
-pub use zk_vm::{MockZkvm, MockCodeCommitment, MockProof};
-pub use da::{TestBlob, TestBlock, TestBlockHeader, TestHash, MockAddress, MockDaSpec};
+pub use zk_vm::{MockCodeCommitment, MockProof, MockZkvm};
